+++ packages/web3-core/test/unit/web3_config.test.ts
@@ -33,6 +33,7 @@ const defaultConfig = {
 		useRpcCallSpecification: false,
 	},
 	handleRevert: false,
+	contractDataInputFill: 'input',
 	maxListenersWarningThreshold: 100,
 	transactionBlockTimeout: 50,
 	transactionConfirmationBlocks: 24,