+++ packages/web3-eth-contract/src/encoding.ts
@@ -187,7 +187,6 @@ export const encodeMethodABI = (
 	deployData?: HexString,
 ) => {
 	const inputLength = Array.isArray(abi.inputs) ? abi.inputs.length : 0;
-
 	if (inputLength !== args.length) {
 		throw new Web3ContractError(
 			`The number of arguments is not matching the methods required number. You need to pass ${inputLength} arguments.`,